[IMP] reactions display

+++ openBlogBack/controllers/postsController.js
@@ -51,6 +51,9 @@ exports.getReactions = async (req, res, next) => {
         const userId = req.params.userId;
         const type = req.params.type;
         const reaction = await postsService.getReactions(postId, userId, type);
+        if(!reaction) {
+            return res.json({ message: "Reaction not found" });
+        }
         res.json(reaction);
     } catch (error) {
         next(new ServerError());
@@ -68,23 +71,16 @@ exports.createPost = async (req, res, next) => {
     }
 }
 
-exports.editPost = async (req, res, next) => {
-    try {
-        const { title, content, postId } = req.body;
-        await postsService.updatePost(postId, title, content);
-        res.json({ message: 'Post updated' });
-    } catch (error) {
-        next(new ServerError());
-    }
-}
-
 exports.addReaction = async (req, res, next) => {
     try {
         const { postId, userId, type } = req.body;
         if(!await usersService.getUserById(userId)) {
             throw createError(404, 'User not found');
         }
-        (await postsService.getReactions(postId, userId, type)) && res.json({ message: "Reaction already exists" });
+        if(await postsService.getReactions(postId, userId, type)){
+            res.json({ message: "Reaction already exists" });
+            return
+        } 
         if (type === 'like' && await postsService.getReactions(postId, userId, 'dislike')) {
             await postsService.removeReaction(postId, userId, 'dislike');
         }
@@ -98,21 +94,23 @@ exports.addReaction = async (req, res, next) => {
     }
 }
 
-exports.deletePost = async (req, res, next) => {
+exports.addSubject = async (req, res, next) => {
     try {
-        const postId = req.params.id;
-        await postsService.deletePost(postId);
-        res.json({ message: 'Post deleted' });
+        const { postId, subjectId } = req.body;
+        await postsService.addSubject(postId, subjectId)
+        res.json({ message: "subject added" })
     } catch (error) {
-        next(new ServerError());
+        next(new 
+            ServerError()
+        );
     }
 }
 
-exports.removeReaction = (req, res, next) => {
+exports.editPost = async (req, res, next) => {
     try {
-        const { postId, userId, type } = req.body;
-        //await postsService.removeReaction(postId, userId, type);
-        res.json({ message: 'Reaction removed' });
+        const { title, content, postId } = req.body;
+        await postsService.updatePost(postId, title, content);
+        res.json({ message: 'Post updated' });
     } catch (error) {
         next(new ServerError());
     }
@@ -128,14 +126,28 @@ exports.incrementNbComments = async (req, res, next) => {
     }
 }
 
-exports.addSubject = async (req, res, next) => {
+exports.deletePost = async (req, res, next) => {
     try {
-        const { postId, subjectId } = req.body;
-        await postsService.addSubject(postId, subjectId)
-        res.json({ message: "subject added" })
+        const postId = req.params.id;
+        await postsService.deletePost(postId);
+        res.json({ message: 'Post deleted' });
     } catch (error) {
-        next(new 
-            ServerError()
-        );
+        next(new ServerError());
+    }
+}
+
+exports.deleteReaction = async (req, res, next) => {
+    try {
+        const postId = req.params.id;
+        const userId = req.params.userId;
+        const type = req.params.type;
+        if(!await postsService.getReactions(postId, userId, type)){
+            res.json({ message: "Reaction not found" });
+            return
+        }
+        await postsService.removeReaction(postId, userId, type);
+        res.json({ message: 'Reaction removed' });
+    } catch (error) {
+        next(new ServerError());
     }
 }
